+++ plugins/org.eclipse.actf.model.ui.editors.ie/src/org/eclipse/actf/model/internal/ui/editors/ie/events/NewWindow2Parameters.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2007 IBM Corporation and Others
+ * Copyright (c) 2007, 2019 IBM Corporation and Others
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -15,9 +15,9 @@ package org.eclipse.actf.model.internal.ui.editors.ie.events;
 
 public interface NewWindow2Parameters {
 
-    public int getBrowserAddress();
+    public long getBrowserAddress();
 
-    public void setBrowserAddress(int address);
+    public void setBrowserAddress(long address);
 
     public boolean getCancel();
