+++ plugins/org.eclipse.actf.model.ui.editors.ie/src/org/eclipse/actf/model/internal/ui/editors/ie/events/impl/NewWindow2ParametersImpl.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2007 IBM Corporation and Others
+ * Copyright (c) 2007, 2019 IBM Corporation and Others
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -27,7 +27,7 @@ public class NewWindow2ParametersImpl extends AbstractEventParameters implements
     public NewWindow2ParametersImpl(OleEvent event) {
         super(event);
         if( Platform.inDebugMode() ) {
-            int     dispBrowser = getBrowserAddress();
+            long     dispBrowser = getBrowserAddress();
             boolean cancel =      getCancel();
             System.out.println("NewWindow2("+dispBrowser+","+cancel+")"); //$NON-NLS-1$ //$NON-NLS-2$ //$NON-NLS-3$
         }
@@ -36,14 +36,14 @@ public class NewWindow2ParametersImpl extends AbstractEventParameters implements
     /* (non-Javadoc)
      * @see org.eclipse.actf.model.ui.editors.ie.events.NewWindow2Parameters#getBrowserAddress()
      */
-    public int getBrowserAddress() {
+    public long getBrowserAddress() {
         return getIntegerByRef(INDEX_Browser);
     }
     
     /* (non-Javadoc)
      * @see org.eclipse.actf.model.ui.editors.ie.events.NewWindow2Parameters#setBrowserAddress(int)
      */
-    public void setBrowserAddress(int address) {
+    public void setBrowserAddress(long address) {
         setIntegerByRef(INDEX_Browser, address);
     }
