+++ plugins/org.eclipse.actf.model.ui.editors.ooo/src/org/eclipse/actf/model/internal/ui/editors/ooo/util/ODFUtils.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2007, 2016 IBM Corporation and Others
+ * Copyright (c) 2007, 2019 IBM Corporation and Others
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -208,13 +208,13 @@ public class ODFUtils {
 		return javaClassLoader;
 	}
 
-	private static String getWindowClass(int hwnd) {
+	private static String getWindowClass(long hwnd) {
 		return WindowUtil.GetWindowClassName(hwnd);
 	}
 
-	private static int getOpenOfficeFrameNum(int hWnd, int frameNum) {
+	private static int getOpenOfficeFrameNum(long hWnd, int frameNum) {
 		int result = frameNum;
-		int hwndChild = WindowUtil.GetChildWindow(hWnd);
+		long hwndChild = WindowUtil.GetChildWindow(hWnd);
 		while (hwndChild != 0) {
 			String winClass = getWindowClass(hwndChild);
 			if (winClass.equals("SALTMPSUBFRAME")) {
@@ -229,7 +229,7 @@ public class ODFUtils {
 
 	public static int getOpenOfficeFrameNum() {
 		int result = 0;
-		int hwndChild = WindowUtil.GetChildWindow(WindowUtil.GetDesktopWindow());
+		long hwndChild = WindowUtil.GetChildWindow(WindowUtil.GetDesktopWindow());
 		// OS.GetWindow (OS.GetDesktopWindow(), OS.GW_CHILD);
 		while (hwndChild != 0) {
 			result = getOpenOfficeFrameNum(hwndChild, result);
