+++ plugins/org.eclipse.actf.util.win32.comclutch/src/org/eclipse/actf/util/win32/comclutch/ComPlugin.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2007 IBM Corporation and Others
+ * Copyright (c) 2007, 2019 IBM Corporation and Others
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -78,7 +78,7 @@ public class ComPlugin extends AbstractUIPlugin {
 	 * @see IKeyHook
 	 */
 	public IKeyHook newKeyHook(IKeyHookListener listener) {
-		int handle = getWorkbench().getActiveWorkbenchWindow().getShell().handle;
+		long handle = getWorkbench().getActiveWorkbenchWindow().getShell().handle;
 		return new KeyHookImpl(listener, handle);
 	}
