+++ plugins/org.eclipse.actf.util.win32/src/org/eclipse/actf/util/internal/win32/AccessibleHilighter.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2007, 2008 IBM Corporation and Others
+ * Copyright (c) 2007, 2019 IBM Corporation and Others
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -40,7 +40,7 @@ public class AccessibleHilighter {
 		if (null != rect) {
 			lastThread = new Thread(new Runnable() {
 				public void run() {
-					int hwndDraw = WindowUtil.GetDesktopWindow();
+					long hwndDraw = WindowUtil.GetDesktopWindow();
 					try {
 						for (int i = 0; i < FLASHING_COLORS.length; i++) {
 							DrawUtil.drawRectangle(hwndDraw, rect,
