+++ plugins/org.eclipse.actf.util.win32/src/org/eclipse/actf/util/internal/win32/DrawUtil.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2007 IBM Corporation and Others
+ * Copyright (c) 2007, 2019 IBM Corporation and Others
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -33,9 +33,9 @@ public class DrawUtil {
 	 * @param color
 	 *            color
 	 */
-	public static void drawRectangle(int hwndDraw, Rectangle rect, int color) {
-		int hDC = OS.GetDCEx(hwndDraw, 0, OS.DCX_CACHE);
-		int oldBrush = OS.SelectObject(hDC, OS.CreateSolidBrush(color));
+	public static void drawRectangle(long hwndDraw, Rectangle rect, int color) {
+		long hDC = OS.GetDCEx(hwndDraw, 0, OS.DCX_CACHE);
+		long oldBrush = OS.SelectObject(hDC, OS.CreateSolidBrush(color));
 		OS.PatBlt(hDC, rect.x, rect.y, rect.width, bandWidth, OS.PATCOPY);
 		OS.PatBlt(hDC, rect.x, rect.y + bandWidth, bandWidth, rect.height
 				- (bandWidth * 2), OS.PATCOPY);
@@ -55,7 +55,7 @@ public class DrawUtil {
 	 * @param rect
 	 *            target rectangle
 	 */
-	public static void redrawRectangle(int hwndDraw, Rectangle rect) {
+	public static void redrawRectangle(long hwndDraw, Rectangle rect) {
 		RECT osRect = new RECT();
 		osRect.left = rect.x - bandWidth;
 		osRect.top = rect.y - bandWidth;
